+++ client/pages/objects/form.tsx
@@ -89,6 +89,7 @@ const ObjectFormPage: NextPage<ObjectFormPageProps> = () => {
                 title={currentPageTitle}
                 description={''}
                 noindex={true}
+                nofollow={true}
                 openGraph={{
                     locale: i18n.language === 'ru' ? 'ru_RU' : 'en_US'
                 }}